✅ Fix some typos, add implementation for ptrvec.c

+++ include/vec.h
@@ -7,17 +7,17 @@
 #include <stdbool.h>
 
 typedef struct Vec {
-    int32_t len;
-    int32_t cap;
-    int32_t elem_size;
+    int64_t len;
+    int64_t cap;
+    int64_t elem_size;
     void *elements;
 } Vec;
 
-Vec vec_create(int32_t elem_size);
-Vec vec_with_cap(int32_t elem_size, int32_t cap);
+Vec vec_create(int64_t elem_size);
+Vec vec_with_cap(int64_t elem_size, int64_t cap);
 void vec_push(Vec *v, void *elem);
-bool vec_get(Vec *v, int32_t i, void *ptr);
-void *vec_get_ptr(Vec *v, int32_t i);
+bool vec_get(Vec *v, int64_t i, void *ptr);
+void *vec_get_ptr(Vec *v, int64_t i);
 void vec_resize(Vec *v);
 void vec_init_zero(Vec *v);
 void vec_free(Vec *v);
